refactor: forEach말구 reduce로변경
