+++ src/api/users.js
@@ -21,3 +21,17 @@ export const fetchUserFolderData = async ({ userId }) => {
     throw new Error("Failed to fetch data");
   }
 };
+
+export const fetchUserLinks = async ({ userId, folderId }) => {
+  const response = await fetch(
+    `https://bootcamp-api.codeit.kr/api/users/${userId}/links${
+      folderId ? `?folderId=${folderId}` : ""
+    }`
+  );
+  if (response.status === 200) {
+    const jsonData = await response.json();
+    return jsonData;
+  } else {
+    throw new Error("Failed to fetch data");
+  }
+};
