+++ src/common/header/Header.js
@@ -3,7 +3,6 @@ import Folder from "../../folder/Folder";
 import FolderSkeleton from "../../components/skeleton/FolderSkeleton";
 import HeaderButton from "../../components/button/HeaderButton";
 export default function Header({ data, isLoading }) {
-  console.log(data);
   if (data) {
     return (
       <div>
