+++ src/hooks/useUserFetch.js
@@ -1,10 +1,10 @@
-import useFetch from "./useFetch";
-import React from "react";
+// import useFetch from "./useFetch";
+// import React from "react";
 
-export default function useUserFetch({ userId }) {
-  const [users, isLoading] = useFetch(
-    `https://bootcamp-api.codeit.kr/api/users/${userId}`
-  );
+// export default function useUserFetch({ userId }) {
+//   const [users, isLoading] = useFetch(
+//     `https://bootcamp-api.codeit.kr/api/users/${userId}`
+//   );
 
-  return [users, isLoading];
-}
+//   return [users, isLoading];
+// }
