+++ src/hooks/userUserFolderFetch.js
@@ -1,8 +1,8 @@
-import useFetch from "./useFetch";
+// import useFetch from "./useFetch";
 
-export default function useUserFolderFetch({ userId }) {
-  const [data, isLoading] = useFetch(
-    `https://bootcamp-api.codeit.kr/api/users/${userId}/folders`
-  );
-  return [data, isLoading];
-}
+// export default function useUserFolderFetch({ userId }) {
+//   const [data, isLoading] = useFetch(
+//     `https://bootcamp-api.codeit.kr/api/users/${userId}/folders`
+//   );
+//   return [data, isLoading];
+// }
